+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/slider/DefaultSliderBuilder.java
@@ -146,6 +146,6 @@ final class DefaultSliderBuilder extends AbstractComponentBuilder<Integer, JSlid
 
 	@Override
 	protected void setInitialValue(JSlider component, Integer initialValue) {
-		component.setValue(initialValue);
+		component.setValue(initialValue == null ? 0 : initialValue);
 	}
 }